yello_interview/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.2.1
djangorestframework==3.14.0
idna==3.4
lxml==4.9.2
markdown-it-py==2.2.0
mdurl==0.1.2
psycopg2==2.9.6
Pygments==2.15.1
pytz==2023.3
requests==2.30.0
rich==13.3.5
sqlparse==0.4.4
urllib3==2.0.2
